split scale tests into separate files
